+++ sample/scripts/prepareConfigsForTesting.sh
@@ -1,0 +1,4 @@
+mv babel.config.js babel.config.orig.js
+mv babel.config.test.js babel.config.js
+mv metro.config.js metro.config.orig.js
+mv metro.config.test.js metro.config.js