+++ sample/src/App.tsx
@@ -17,7 +17,6 @@ import EndToEndTestsScreen from './screens/EndToEndTestsScreen';
 import ReduxScreen from './screens/ReduxScreen';
 
 import {store} from './reduxApp';
-import {version as packageVersion} from '../../package.json';
 import {SENTRY_INTERNAL_DSN} from './dsn';
 
 const reactNavigationV5Instrumentation = new Sentry.ReactNavigationV5Instrumentation(
@@ -60,8 +59,8 @@ Sentry.init({
   tracesSampleRate: 1.0,
   // Sets the `release` and `dist` on Sentry events. Make sure this matches EXACTLY with the values on your sourcemaps
   // otherwise they will not work.
-  release: packageVersion,
-  dist: `${packageVersion}.0`,
+  release: '1.2.3',
+  dist: `1.2.3.0`,
 });
 
 const Stack = createStackNavigator();